Make curly braces always single-colored
Before this commit, if the optional “color” keyword was not given,
then the curly brace would have several, seemingly random colors.
This two-line fix ensures that a curly brace has always a single color,
either given as the keyword argument or determined automatically.

A new example script making use of this is added.